2019-12-03 update

+++ lib/common.lib.php
@@ -2405,24 +2405,59 @@ class html_process {
     function run()
     {
         global $config, $g5, $member;
+        global $device,$mac,$app,$app2;
 
-        // 현재접속자 처리
-        $tmp_sql = " select count(*) as cnt from {$g5['login_table']} where lo_ip = '{$_SERVER['REMOTE_ADDR']}' ";
-        $tmp_row = sql_fetch($tmp_sql);
+        if($member["mb_id"]) {
+            if ($app || $app2) {
+                //앱일경우
+                if ($device == "") {
+                    $device = get_session("device");
+                }
+                if ($mac == "") {
+                    $mac = get_session("mac");
+                }
 
-        if ($tmp_row['cnt']) {
-            $tmp_sql = " update {$g5['login_table']} set mb_id = '{$member['mb_id']}', lo_datetime = '".G5_TIME_YMDHIS."', lo_location = '{$g5['lo_location']}', lo_url = '{$g5['lo_url']}' where lo_ip = '{$_SERVER['REMOTE_ADDR']}' ";
-            sql_query($tmp_sql, FALSE);
-        } else {
-            $tmp_sql = " insert into {$g5['login_table']} ( lo_ip, mb_id, lo_datetime, lo_location, lo_url ) values ( '{$_SERVER['REMOTE_ADDR']}', '{$member['mb_id']}', '".G5_TIME_YMDHIS."', '{$g5['lo_location']}',  '{$g5['lo_url']}' ) ";
-            sql_query($tmp_sql, FALSE);
+                $tmp_sql = " select count(*) as cnt from {$g5['login_table']} where lo_device = '{$device}' and lo_device_id = '{$mac}' and (lo_device <> '' and lo_device_id <> '')";
+                $tmp_row = sql_fetch($tmp_sql);
 
-            // 시간이 지난 접속은 삭제한다
-            sql_query(" delete from {$g5['login_table']} where lo_datetime < '".date("Y-m-d H:i:s", G5_SERVER_TIME - (60 * $config['cf_login_minutes']))."' ");
+                if ($tmp_row['cnt']) {
+                    //로그인 정보 업데이트
+                    $tmp_sql = " update {$g5['login_table']} set mb_id = '{$member['mb_id']}', lo_datetime = '" . G5_TIME_YMDHIS . "', lo_location = '{$g5['lo_location']}', lo_url = '{$g5['lo_url']}' , lo_device = '{$device}', lo_device_id = '{$mac}' where lo_device = '{$device}' and lo_device_id = '{$mac}'";
+                    sql_query($tmp_sql, FALSE);
+                } else {
+                    $tmp_sql = " insert into {$g5['login_table']} ( lo_ip, mb_id, lo_datetime, lo_location, lo_url, lo_device, lo_device_id) values ( '{$_SERVER['REMOTE_ADDR']}', '{$member['mb_id']}', '" . G5_TIME_YMDHIS . "', '{$g5['lo_location']}',  '{$g5['lo_url']}' , '{$device}', '{$mac}') ";
+                    sql_query($tmp_sql, FALSE);
 
-            // 부담(overhead)이 있다면 테이블 최적화
-            //$row = sql_fetch(" SHOW TABLE STATUS FROM `$mysql_db` LIKE '$g5['login_table']' ");
-            //if ($row['Data_free'] > 0) sql_query(" OPTIMIZE TABLE $g5['login_table'] ");
+                    // 시간이 지난 접속은 삭제한다
+                    sql_query(" delete from {$g5['login_table']} where lo_datetime < '" . date("Y-m-d H:i:s", G5_SERVER_TIME - (60 * $config['cf_login_minutes'])) . "' ");
+                }
+            } else {
+                //현재 접속 아이피 확인
+/*                $chksql = "select * from `mydevice` where mb_id = '{$member["mb_id"]}'";
+                echo $chksql;
+                $chkrow = sql_fetch($chksql);
+                if($chkrow==null){
+
+                }*/
+
+                // 현재접속자 처리
+                $tmp_sql = " select count(*) as cnt from {$g5['login_table']} where lo_ip = '{$_SERVER['REMOTE_ADDR']}' and mb_id = '{$member["mb_id"]}' and isnull(lo_device) and isnull(lo_device_id) ";
+                $tmp_row = sql_fetch($tmp_sql);
+                if ($tmp_row['cnt']) {
+                    $tmp_sql = " update {$g5['login_table']} set mb_id = '{$member['mb_id']}', lo_datetime = '" . G5_TIME_YMDHIS . "', lo_location = '{$g5['lo_location']}', lo_url = '{$g5['lo_url']}', lo_ip = '{$_SERVER["REMOTE_ADDR"]}' where mb_id = '{$member["mb_id"]}' and isnull(lo_device) and isnull(lo_device_id)";
+                    sql_query($tmp_sql, FALSE);
+                } else {
+                    $tmp_sql = " insert into {$g5['login_table']} ( lo_ip, mb_id, lo_datetime, lo_location, lo_url) values ( '{$_SERVER['REMOTE_ADDR']}', '{$member['mb_id']}', '" . G5_TIME_YMDHIS . "', '{$g5['lo_location']}',  '{$g5['lo_url']}' ) ";
+                    sql_query($tmp_sql, FALSE);
+
+                    // 시간이 지난 접속은 삭제한다
+                    sql_query(" delete from {$g5['login_table']} where lo_datetime < '" . date("Y-m-d H:i:s", G5_SERVER_TIME - (60 * $config['cf_login_minutes'])) . "' ");
+
+                    // 부담(overhead)이 있다면 테이블 최적화
+                    //$row = sql_fetch(" SHOW TABLE STATUS FROM `$mysql_db` LIKE '$g5['login_table']' ");
+                    //if ($row['Data_free'] > 0) sql_query(" OPTIMIZE TABLE $g5['login_table'] ");
+                }
+            }
         }
 
         $buffer = ob_get_contents();
@@ -2655,8 +2690,8 @@ function certify_count_check($mb_id, $type)
             break;
     }
 
-    if((int)$row['cnt'] >= (int)$config['cf_cert_limit'])
-        alert_close('오늘 '.$cert.' 본인확인을 '.$row['cnt'].'회 이용하셔서 더 이상 이용할 수 없습니다.');
+    //if((int)$row['cnt'] >= (int)$config['cf_cert_limit'])
+        //alert_close('오늘 '.$cert.' 본인확인을 '.$row['cnt'].'회 이용하셔서 더 이상 이용할 수 없습니다.');
 }
 
 // 1:1문의 설정로드
